requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alabaster==0.7.16
albucore==0.0.23
albumentations==2.0.0
alembic==1.14.0
annotated-types==0.7.0
arrow==1.3.0
astroid==3.3.8
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncssh==2.19.0
atomicwrites==1.4.1
attrs==24.3.0
autopep8==2.0.4
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.10.0
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpickle==3.1.1
cmake==3.31.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cookiecutter==2.6.0
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
deepface==0.0.93
defusedxml==0.7.1
Deprecated==1.2.15
diff-match-patch==20241021
dill==0.3.9
distlib==0.3.9
dlib @ file:///D:/bld/dlib-split_1726093520894/work
dnspython==2.7.0
docopt==0.6.2
docstring-to-markdown==0.15
docutils==0.21.2
dominate==2.9.1
dulwich==0.22.7
easydict==1.13
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.7.0
flake8==7.1.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
gunicorn==23.0.0
h5py==3.12.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.0.0
insightface==0.7.3
installer==0.7.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.18.1
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
mtcnn==0.1.1
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy @ file:///D:/bld/numpy_1732314314278/work/dist/numpy-2.0.2-cp39-cp39-win_amd64.whl#sha256=8824ea87142626db7bbf79bdbcbd5b3a2fd77d7570bd62b4f705100e42b651e1
numpydoc==1.8.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
prettytable==3.12.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.5.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.3
pylint-venv==3.0.4
pyls-spyder==0.4.0
pymongo==4.10.1
PyNaCl==1.5.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.2
pyuca==1.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.2
RapidFuzz==3.11.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==13.9.4
rope==1.13.0
rpds-py==0.22.3
Rtree==1.3.0
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.3
spyder-kernels==3.0.2
SQLAlchemy==2.0.37
stack-data==0.6.3
stringzilla==3.11.3
superqt==0.7.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
text-unidecode==1.3
textdistance==4.6.3
tf_keras==2.18.0
threadpoolctl==3.5.0
three-merge==0.1.1
tifffile==2024.8.30
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.1.15.22
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.3.61
ultralytics-thop==2.0.13
urllib3==2.3.0
virtualenv==20.29.1
visitor==0.1.3
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whatthepatch==1.0.7
wincertstore==0.2.1
wrapt==1.17.1
WTForms==3.2.1
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
